feat(magazine): introduce MagazineList component for magazine display
- Created a new MagazineList component to encapsulate the magazine listing functionality, including loading states, latest issue display, and pagination.
- Updated the Magazine.vue page to utilize the new MagazineList component, streamlining the code and enhancing maintainability.
- Added a new route for accessing magazines directly, improving navigation within the application.

Overall: This commit enhances the magazine viewing experience by modularizing the component structure and improving the organization of magazine-related features.
